+++ CHANGELOG.md
@@ -8,6 +8,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Changed
+- On instantiation of the RestClient no authentication is done automatically anymore
 - Incident class accepts as a priority only a Priority enum
 - Incident class must be instantiated with a step number
 - Step number in Incident class is always an integer