+++ docs/changelog.rst
@@ -15,6 +15,7 @@ Changed
 ^^^^^^^
 
 
+* On instantiation of the RestClient no authentication is done automatically anymore
 * Incident class accepts as a priority only a Priority enum
 * Incident class must be instantiated with a step number
 * Step number in Incident class is always an integer